refactor: split consts.h into consts.h and types.h
